+++ ui.apps/src/main/content/jcr_root/apps/acs-commons/components/utilities/contentsync/clientlibs/js/app.js
@@ -21,12 +21,12 @@
     var ui = $(window).adaptTo("foundation-ui");
     
     $(window).adaptTo("foundation-registry").register("foundation.collection.action.action", {
-        name: "acs-commons.dashboard.task.delete",
+        name: "acs-commons.contentsync.host.delete",
         handler: function(name, el, config, collection, selections) {	
             if (selections.length === 1) {
 				var card = $(selections).find("coral-card");
                 if(card.length > 0) {
-					var taskName = card.data("host");
+                    var taskName = card.data("properties").name;
                     var message = "You are going to delete the following item:" +
                     "<p><b>" + taskName + "</b></p>";
 
@@ -59,7 +59,7 @@
         handler: function(name, el, config, collection, selections) {
             var dlg = document.getElementById("modalConfigureHost");
             var form = dlg.querySelector("form");
-            form.action = "/etc/replication/agents.author/*";
+            form.action = "/var/acs-commons/contentsync/hosts/*";
             form.reset();
             dlg.show();
         }
@@ -90,7 +90,7 @@
 
 	document.addEventListener("DOMContentLoaded", function() {
         var dlg = document.getElementById("modalConfigureHost");
-    	dlg.querySelector("coral-select[name=\"./authType\"]")
+    	if(dlg) dlg.querySelector("coral-select[name=\"./authType\"]")
             .addEventListener("change", function(e) {
             	var authType = e.target.value;
             	dlg.querySelectorAll(".list-option-showhide-target").forEach(function(c){
@@ -103,8 +103,22 @@
         	}	
         );
         
+        var workflowModelSelect = $("coral-select[name=workflowModel]");
+        if(workflowModelSelect) {
+	    	workflowModelSelect.get(0).value = $("input[name=defaultWorkflow]").get(0).value;
+        }
     });
     
 
-    
+    $(window).adaptTo("foundation-registry").register("foundation.form.response.ui.error", {
+        name: "foundation.default",
+        handler: function(a, g, e, b, d) {
+            a = "Error";
+            d = "Fail to submit the form. Make sure you have permissions to run content sync.";
+            ui.alert(a, d, "error");
+        }
+    });
+
+
+
 })(document, Granite.$);